:recycle: CHORE
Modified and splited types location
