+++ src/containers/MessageContainer.tsx
@@ -1,15 +1,14 @@
 import React, { useEffect } from 'react'
-import { useAppSelector } from '../hooks'
-import { Message, MessageType } from '../state/types'
 import styled from 'styled-components'
 
-import MessageBubbleComponent from '../component/MessageBubbleComponent'
+import { useAppSelector } from '../hooks'
+import { Message } from "../types"
+import MessageBubbleComponent from '../components/MessageBubbleComponent'
 
 const Container = styled.div`
   width: 90vw;
-  margin-top: 15vh;
-  padding: 5vh 0 5vh 0;
-  height: 75vh;
+  margin-top: 20vh;
+  height: 65vh;
   z-index: 0;
   max-width: 45rem;
   position: absolute;
