+++ src/state/state.types.ts
@@ -1,0 +1,10 @@
+import { Message, Client } from "../types";
+
+export type MessageState = {
+  messages: Message[]
+};
+
+export type ClientState = {
+  currentClient?: Client,
+  clients: Client[]
+}
